client system

+++ data/system_location.cpp
@@ -5,8 +5,8 @@ namespace planetary_mistral {
         system_location::system_location(
                 const int altitude,
                 const QString& landscape_key,
-                const double latitude,
-                const double longitude,
+                const long double latitude,
+                const long double longitude,
                 const QString& name,
                 const QString& planet,
                 const QString& region,
@@ -25,8 +25,8 @@ namespace planetary_mistral {
         system_location::system_location(
                 const int altitude,
                 QString&& landscape_key,
-                const double latitude,
-                const double longitude,
+                const long double latitude,
+                const long double longitude,
                 QString&& name,
                 QString&& planet,
                 QString&& region,
@@ -45,8 +45,8 @@ namespace planetary_mistral {
         system_location::system_location(
                 const int altitude,
                 const std::u16string& landscape_key,
-                const double latitude,
-                const double longitude,
+                const long double latitude,
+                const long double longitude,
                 const std::u16string& name,
                 const std::u16string& planet,
                 const std::u16string& region,
